+++ Implementation/data_structures/art/node/node48.cpp
@@ -49,7 +49,7 @@ namespace art
         const uint8_t from_key = from >> offset & 0xFF;
         const uint8_t to_key = to >> offset & 0xFF;
 
-        for (uint8_t i = from_key; i <= to_key; ++i)
+        for (uint16_t i = from_key; i <= to_key; ++i)
         {
             if (keys_[i] == free_marker_) continue;
